requirements.txt
fastapi
uvicorn[standard]
pandas
requests
python-dotenv
jinja2